🐛 최근검색어 삭제 이슈로 기존코드로 돌아감
